+++ src/main/groovy/info/solidsoft/gradle/pitest/PitestTask.groovy
@@ -18,7 +18,9 @@ package info.solidsoft.gradle.pitest
 import groovy.transform.CompileStatic
 import groovy.transform.PackageScope
 import org.gradle.api.Incubating
+import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.FileCollection
+import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.MapProperty
@@ -38,17 +40,17 @@ import org.gradle.api.tasks.options.Option
  * Gradle task implementation for Pitest.
  */
 @CompileStatic
+@SuppressWarnings("UnstableApiUsage")   //@Option
 class PitestTask extends JavaExec {
 
     @Input
     @Optional
     final Property<String> testPlugin
 
-//    //ClassNotFoundException: org.gradle.api.file.FileSystemLocationProperty in Gradle <5.6 due to super interface of RegularFileProperty
-//    RegularFileProperty reportDir
-    @Incubating //will be replaced with RegularFileProperty when switched to Gradle 5.6+
+    //ClassNotFoundException: org.gradle.api.file.FileSystemLocationProperty in Gradle <5.6 due to super interface of RegularFileProperty
+    //See: https://github.com/gradle/gradle/issues/10953 - on the other hand deprecation warnings in Gradle 6.x with regular File
     @OutputDirectory
-    File reportDir
+    final DirectoryProperty reportDir
 
     @Input
     final SetProperty<String> targetClasses
@@ -156,7 +158,6 @@ class PitestTask extends JavaExec {
     final Property<Boolean> useAdditionalClasspathFile
 
     @Input
-    @OutputFile
     File additionalClasspathFile
 
     @InputFiles
@@ -164,16 +165,17 @@ class PitestTask extends JavaExec {
 
     @Input
     @Optional
-    File historyInputLocation
+    final RegularFileProperty historyInputLocation
 
     @OutputFile
     @Optional
-    File historyOutputLocation
+    final RegularFileProperty historyOutputLocation
 
     @Input
     @Optional
     final Property<Boolean> enableDefaultIncrementalAnalysis
 
+    //TODO: Switch to RegularFileProperty once https://github.com/gradle/gradle/issues/12351 is solved
     @Input
     File defaultFileForHistoryData
 
@@ -195,7 +197,7 @@ class PitestTask extends JavaExec {
 
     @Input
     @Optional
-    File jvmPath
+    final RegularFileProperty jvmPath
 
     @Input
     @Optional
@@ -223,17 +225,21 @@ class PitestTask extends JavaExec {
 
     @Incubating
     @Option(option = "additionalFeatures", description = "Additional PIT features to be appended to those placed in configuration")
+    @Input
+    @Optional
     List<String> additionalFeatures //ListProperty<String> cannot be used with @Option - https://github.com/gradle/gradle/issues/10517
 
     @Incubating
     @Option(option = "targetTests", description = "Tests classes to use. Overrides 'testClasses' defined in configuration")
+    @Input
+    @Optional
     List<String> overriddenTargetTests  //should be Set<String> or SetProperty but it's not supported in Gradle as of 5.6.1
 
     PitestTask() {
         ObjectFactory of = project.objects
 
         testPlugin = of.property(String)
-//        reportDir = of.fileProperty()
+        reportDir = of.directoryProperty()
         targetClasses = of.setProperty(String)
         targetTests = of.setProperty(String)
         dependencyDistance = of.property(Integer)
@@ -259,15 +265,15 @@ class PitestTask extends JavaExec {
 //        sourceDirs = of.fileProperty()
         detectInlinedCode = of.property(Boolean)
         timestampedReports = of.property(Boolean)
-//        historyInputLocation = of.fileProperty()
-//        historyOutputLocation = of.fileProperty()
+        historyInputLocation = of.fileProperty()
+        historyOutputLocation = of.fileProperty()
         enableDefaultIncrementalAnalysis = of.property(Boolean)
 //        defaultFileForHistoryData = of.fileProperty()
         mutationThreshold = of.property(Integer)
         coverageThreshold = of.property(Integer)
         mutationEngine = of.property(String)
         exportLineCoverage = of.property(Boolean)
-//        jvmPath = of.fileProperty()
+        jvmPath = of.fileProperty()
         mainProcessJvmArgs = of.listProperty(String)
 //        mutableCodePaths = of.setProperty(File)
         pluginConfiguration = of.mapProperty(String, String)
@@ -299,8 +305,7 @@ class PitestTask extends JavaExec {
     Map<String, String> createTaskArgumentMap() {
         Map<String, String> map = [:]
         map['testPlugin'] = testPlugin.getOrNull()
-//        map['reportDir'] = reportDir.getOrNull()?.toString()
-        map['reportDir'] = getReportDir().toString()
+        map['reportDir'] = reportDir.getOrNull()?.toString()
         map['targetClasses'] = targetClasses.get().join(',')
         map['targetTests'] = overriddenTargetTests ? overriddenTargetTests.join(',') : optionalCollectionAsString(targetTests)
         map['dependencyDistance'] = optionalPropertyAsString(dependencyDistance)
@@ -332,7 +337,7 @@ class PitestTask extends JavaExec {
         map['mutationEngine'] = mutationEngine.getOrNull()
         map['exportLineCoverage'] = optionalPropertyAsString(exportLineCoverage)
         map['includeLaunchClasspath'] = Boolean.FALSE.toString()   //code to analyse is passed via classPath
-        map['jvmPath'] = getJvmPath()?.path
+        map['jvmPath'] = getJvmPath()?.getOrNull()?.toString()
         map['maxSurviving'] = optionalPropertyAsString(maxSurviving)
         map['useClasspathJar'] = optionalPropertyAsString(useClasspathJar)
         map['features'] = (features.getOrElse([]) + (additionalFeatures?: []))?.join(',')
@@ -361,11 +366,11 @@ class PitestTask extends JavaExec {
 
     private Map<String, String> prepareMapWithIncrementalAnalysisConfiguration() {
         if (enableDefaultIncrementalAnalysis.getOrNull()) {
-            return [historyInputLocation : getHistoryInputLocation()?.path ?: getDefaultFileForHistoryData().path,
-                    historyOutputLocation: getHistoryOutputLocation()?.path ?: getDefaultFileForHistoryData().path]
+            return [historyInputLocation : getHistoryInputLocation()?.getOrNull()?.asFile?.path ?: getDefaultFileForHistoryData().path,
+                    historyOutputLocation: getHistoryOutputLocation()?.getOrNull()?.asFile?.path ?: getDefaultFileForHistoryData().path]
         } else {
-            return [historyInputLocation: getHistoryInputLocation()?.path,
-                    historyOutputLocation: getHistoryOutputLocation()?.path]
+            return [historyInputLocation: getHistoryInputLocation()?.getOrNull()?.asFile?.path,
+                    historyOutputLocation: getHistoryOutputLocation()?.getOrNull()?.asFile?.path]
         }
     }